Add final pass/fail check for merge eligibility
