+++ script/github-actions/verify-merge-eligibility.js
@@ -1,0 +1,21 @@
+const core = require('@actions/core');
+
+const testsBlockingMerge = process.env.TESTS_BLOCKING_MERGE
+  ? JSON.parse(process.env.TESTS_BLOCKING_MERGE)
+  : [];
+
+const e2eAnnotations = process.env['e2e-annotations-json']
+  ? JSON.parse(process.env['e2e-annotations-json'])
+  : [];
+
+const unitTestAnnotations = process.env['unit-test-annotations-json']
+  ? JSON.parse(process.env['unit-test-annotations-json'])
+  : [];
+
+if (e2eAnnotations.length > 0 || unitTestAnnotations.length > 0) {
+  core.setFailed(
+    `This PR contains code to test specs that are disallowed for flakiness. Currently there is a grace period, beginning on October 4, to allow time for tests to be corrected. If not corrected by November 6, this application will be blocked from merging code changes. The file paths causing this status are: \n ${testsBlockingMerge.join(
+      '\n',
+    )}`,
+  );
+}
